### 2019-07-15 #1568 @whedon
[PRE REVIEW]: G3F: Global, Multidimensional Spectral Regression Analysis
**Submitting author:** @dap-biospec (<a href="http://orcid.org/0000-0003-4625-0323">Denis Proshlyakov</a>)
**Repository:** <a href="https://github.com/dap-biospec/G3F" target ="_blank">https://github.com/dap-biospec/G3F</a>
**Version:** v1.0a4
