### 2019-07-15 #1568 @arfon
:wave: @dap-biospec - this PR should fix your paper compilation: https://github.com/dap-biospec/G3F/pull/1
