### 2019-07-15 #1568 @dap-biospec
@arfon @kyleniemeyer Thank you for patching up the paper!

Below are three suggested reviewers. I have no affiliation with any of them.

Christopher Neese: physical chemistry, chemical physics, https://electroscience.osu.edu/people/neese.6

Tim Causgrove: chemistry, protein kinetics and thermodynamics http://faculty.tamucc.edu/tcausgrove/Causgrove_Home.html

Tony Withers: geochemistry; equilibrium and dynamics; spectroscopy; - http://www.bgi.uni-bayreuth.de/?page=4&lng=en&mode=s&id=337
