### 2019-07-18 #1568 @kyleniemeyer
@dap-biospec could you merge @arfon's PR for your paper fix, so that we can compile it here?
