+++ paper.bib
@@ -21,8 +21,8 @@
 	Journal = {Analytical Chemistry},
 	Month = july,
 	Title = {{FTIR Spectro-Voltammetry and Quantitative Modeling of Analytes in Kinetically Constrained Redox Mixtures}},
-	Year = 2019
-	adsurl = {https://pubs.acs.org/doi/10.1021/acs.analchem.9b00859},
+	Year = 2019,
+	adsurl = {https://pubs.acs.org/doi/10.1021/acs.analchem.9b00859}
 }
 
 @article{John:2019b,
