+++ paper.bib
@@ -1,4 +1,4 @@
-@article{Gryzka:2010,
+@article{Grzyska:2010,
 	Author = {{Grzyska}, P. ~K. and {Appelman}, E. ~H and {Hausinger}, R. ~P and {Proshlyakov}, D. ~A.},
 	Journal = {PNAS},
 	Month = jan,
