### 2019-07-21 16d3fe3
Update paper.md
