+++ README.md
@@ -33,7 +33,7 @@ Analysis -> Global 3D Spectral Regression -> Control Panel
 
 For further details on G3F structure and function, please see:<br/>
 [G3F API](https://github.com/dap-biospec/G3F/blob/master/Docs/G3F_API.md)<br/>
-[G3F Manual](https://github.com/dap-biospec/G3F/blob/master/G3F_Manual.md)
+[G3F Manual](https://github.com/dap-biospec/G3F/blob/master/Docs/G3F_Manual.md)
 
 A [Demo Experiment](https://github.com/dap-biospec/G3F/blob/master/Demo) is included with this package. The tutorial is in the [Demo Experiment](https://github.com/dap-biospec/G3F/blob/master/G3F_Manual.md#demo-experiment) section of the manual.
 
