### 2019-07-23 04b0ae0
Update G3F_Manual.md
