### 2019-07-23 ed420f1
Update G3F_Manual.md
