+++ Docs/G3F_Manual.md
@@ -44,21 +44,21 @@ The overall data structure of the datasets handled by G3F is shown below:
 
 Data sets consist of &quot;layers&quot; of matrices with &quot;row&quot;, &quot;column&quot; and &quot;layer&quot; calibrations. The 2D layers of this data structure are shown conceptually below:
 
-![alt text](https://github.com/dap-biospec/G3F/blob/master/G3FManualPictures/DataStructure_Simple.JPG)
+![alt text](https://github.com/dap-biospec/G3F/blob/master/Docs/G3FManualPictures/DataStructure_Simple.JPG)
 
 The dimensionality of this data structure is further expanded with Layer-Column (LCP2D) and Layer-Row (LRP2D) dimensions, as shown below:
 
-![alt text](https://github.com/dap-biospec/G3F/blob/master/G3FManualPictures/DataStructure_2.png)
+![alt text](https://github.com/dap-biospec/G3F/blob/master/Docs/G3FManualPictures/DataStructure_2.png)
  
  An alternative depiction of this data structure with emphasis on the roles of global parameters and local parameters is shown below:
 
- ![alt text](https://github.com/dap-biospec/G3F/blob/master/G3FManualPictures/DataStructure_3.jpg)
+ ![alt text](https://github.com/dap-biospec/G3F/blob/master/Docs/G3FManualPictures/DataStructure_3.jpg)
  
 Row, column, and layer locals are calculated parameters, whereas global parameters are parameters which remain consistent throughout the data set.
 
 ### Summary of G3F User Interface
 
-![alt text](https://github.com/dap-biospec/G3F/blob/master/G3FManualPictures/ControlPanel.png)
+![alt text](https://github.com/dap-biospec/G3F/blob/master/Docs/G3FManualPictures/ControlPanel.png)
 
 The main control panel of the G3F macro is split into several different sections, each of which are detailed below:
 
@@ -96,7 +96,7 @@ This section covers the input data parameters.
 
 Masks are used to block out signals from fitting. Sections of data may be &quot;masked&quot; and therefore not considered for the fit. An example of this concept is spectroscopic data when fitting a baseline – peak signals must be excluded from the baseline fit in order to preserve meaningful data through the baseline fit.
 
-![alt text](https://github.com/dap-biospec/G3F/blob/master/G3FManualPictures/Mask_Example.jpg)
+![alt text](https://github.com/dap-biospec/G3F/blob/master/Docs/G3FManualPictures/Mask_Example.jpg)
 
 - Z(Col): This allows the user to specify a mask wave for the column dimension of data sets
 
@@ -330,7 +330,7 @@ This procedure is a simple Lorentzian function for several isotopes with local v
 
 On the **Method** section on the Control Panel, the user may select a pre-compiled fitting function on the drop-down menu **Fit to function** as shown below:
 
-![alt text](https://github.com/dap-biospec/G3F/blob/master/G3FManualPictures/LoadFunction.png)
+![alt text](https://github.com/dap-biospec/G3F/blob/master/Docs/G3FManualPictures/LoadFunction.png)
 
 The function selected in the **Fit to function** dialog will be used to fit experimental data.
 
@@ -338,11 +338,11 @@ The function selected in the **Fit to function** dialog will be used to fit expe
 
 Three-dimensional datasets may be loaded as a 3D matrix into IgorPro and selected using the **Dataset** dropdown menu as shown below:
 
-![alt text](https://github.com/dap-biospec/G3F/blob/master/G3FManualPictures/LoadData_1.png) 
[…]
