+++ Demo/readme.md
@@ -8,7 +8,7 @@ Download G3F as described in the Installation section.
 
 Download folder and save as a sibling folder of the G3F folder.
 
-Open Demo\_G3F.pxp
+Open Demo_G3F.pxp
 
 Experiment should load G3F code automatically. If G3F\_main.ipf cannot be found, IgorPro will present a dialog that will allow to navigate to the G3F folder.
 
@@ -16,7 +16,7 @@ Demo experiment will also load the SpecEchem\_4Spec\_ForOxd\_2.ipf procedure fro
 
 The main control panel should look like this:
 
-![alt text](https://github.com/dap-biospec/G3F/blob/master/DemoPictures/Pic_1_Demo.png)
+![alt text](https://github.com/dap-biospec/G3F/blob/master/Demo/DemoPictures/Pic_1_Demo.png)
 
 The control panel should be set up for package testing.
 
@@ -30,11 +30,11 @@ Go to the **Fit to function** drop-down menu. Select &quot;SpecEChem_4Spec_ForOx
 
 The global parameters displayed on the control panel correspond to the redox potentials and electron transfer coefficients of each mediator.
 
-![alt text](https://github.com/dap-biospec/G3F/blob/master/DemoPictures/Pic_2_Demo.png)
+![alt text](https://github.com/dap-biospec/G3F/blob/master/Demo/DemoPictures/Pic_2_Demo.png)
 
 As there are 2 mediators and myoglobin present, the data must be deconvoluted to isolate the spectrum and Nernstian profile of myoglobin. To do this, four local parameters are set (methylene green has two electron transfer steps), each corresponding to a spectrum specific to the analyte.
 
-![alt text](https://github.com/dap-biospec/G3F/blob/master/DemoPictures/Pic_3_Demo.png)
+![alt text](https://github.com/dap-biospec/G3F/blob/master/Demo/DemoPictures/Pic_3_Demo.png)
 
 In this experiment, all baselines are set to zero (Col(Z) Locals). These are not used in this simulation.
 
@@ -42,7 +42,7 @@ Next, holds must be set on known parameters. The electron transfer coefficients 
 
 Go to the **Matrix Constraints** button.
 
-![alt text](https://github.com/dap-biospec/G3F/blob/master/DemoPictures/Pic_4_Demo.png)
+![alt text](https://github.com/dap-biospec/G3F/blob/master/Demo/DemoPictures/Pic_4_Demo.png)
 
 Hold all variables (K0 through K5) by checking the **Hold** field. These correspond to the electron transfer coefficients and redox potentials of the mediators – all of which are known. The last two fields correspond to the electron transfer coefficient and redox potential of myoglobin. When finished, press **Done**.
 
@@ -54,11 +54,11 @@ Now that the simulation has been performed, the separate spectra can be plotted 
 
 Go to the Data Browser field and select Oxidation\_RowLoc
 
-![alt text](https://github.com/dap-biospec/G3F/blob/master/DemoPictures/Pic_5_Demo.png)
+![alt text](https://github.com/dap-biospec/G3F/blob/master/Demo/DemoPictures/Pic_5_Demo.png)
 
 Clicking on this brings up a set of data:
 
-![alt text](https://github.com/dap-biospec/G3F/blob/master/DemoPictures/Pic_6_Demo.png)
+![alt text](https://github.com/dap-biospec/G3F/blob/master/Demo/DemoPictures/Pic_6_Demo.png)
 
 This fitted data set corresponds to each individual spectrum of each mediator and myoglobin. To plot this data for easy comparison, go to
 
@@ -66,15 +66,15 @@ Windows ->New Graph…
 
[…]
