### 2019-07-23 456586a
Update G3F_Manual.md
