### 2019-07-25 20ca9c1
Update G3F_Manual.md
