### 2019-07-30 c71086f
Update of demo and performance testing
-Demo experiment now includes three separate modes of fitting. Documentation has been updated accordingly.
-Included separate performance testing experiment and documentaion on verifying fitttign results under pre-configured conditions.
