+++ Demo/readme.md
@@ -1,24 +1,32 @@
 ### Demo Experiment
 
-To demonstrate the efficacy of G3F, an example NPSV experiment is provided with the G3F package.  [Demo Experiment](https://github.com/dap-biospec/G3F/tree/master/Demo)This demo models the extraction of the redox profile of myoglobin with different spectra taken at different applied potentials. In addition to different applied potentials, different compositions of 2 different electrochemical mediators (methylene green, thioninie acetate) are present in this data. In order to analyze the redox profiles of myoglobin alone, the data must be deconvoluted.
+To demonstrate the efficacy of G3F, an example NPSV experiment is provided with the G3F package.  [Demo Experiment](https://github.com/dap-biospec/G3F/tree/master/Demo) This demo models the extraction of the redox profile of myoglobin with different spectra taken at different applied potentials. In addition to different applied potentials, different compositions of 2 different electrochemical mediators (methylene green, thioninie acetate) are present in this data. In order to analyze the redox profiles of myoglobin alone, the data must be deconvoluted. This analysis is demonstrated in three separate ways in order to familiarize the user with the flexibility of the G3F package.
+
+- [Example #1: Direct Fitting](#direct-fitting)
+- [Example #2: Process Fitting](#process-fitting)
+- [Example #3: Calculating Fits Using Local Variables](#calculating-fits-using-local-variables)
 
 **Installing and Opening:**
 
-Download G3F as described in the Installation section.
+Download G3F as described in the [Installation section](https://github.com/dap-biospec/G3F/blob/master/Docs/G3F_Manual.md#installation).
 
 Download folder and save as a sibling folder of the G3F folder.
 
 Open Demo_G3F.pxp
 
-Experiment should load G3F code automatically. If G3F\_main.ipf cannot be found, IgorPro will present a dialog that will allow to navigate to the G3F folder.
+Experiment should load G3F code automatically. If G3F_main.ipf cannot be found, IgorPro will present a dialog that will allow the user to navigate to the G3F folder.
 
-Demo experiment will also load the SpecEchem\_4Spec\_ForOxd\_2.ipf procedure from the Demo folder with the example of user-supplied fitting function. 
+Demo experiment will also load the SpecEchem_4Spec_ForOxd_2.ipf procedure from the Demo folder with the example of user-supplied fitting function. 
 
-The main control panel should look like this:
+Open the main control panel, and type in the values for the global variables:
 
 ![alt text](https://github.com/dap-biospec/G3F/blob/master/Demo/DemoPictures/Pic_1_Demo.png)
 
 The control panel should be set up for package testing.
+
+### **Direct Fitting**
+
+In this example, the data will be fitted directly using a fitting function and global variables for known parameters of analytes (ie: standard reduction potentials, number of electrons transferred). Individual spectra corresponding to each analyte and the Nernstian profile of myoglobin are calculated using local variables.
 
 **Loading Raw Data and Fitting Function**
 
@@ -94,4 +102,199 @@ By formatting the raw data as dots and keeping the fit solid, a graph should be 
 
 From these traces, the redox potential of myoglobin can be determined. The multidimensional data has been successfully deconvoluted.
 
- 
+### **Process Fitting**
+
[…]
