+++ Docs/G3F_Manual.md
@@ -8,7 +8,7 @@ The G3F (Global 3-dimensional Fit) is a simulation and fitting package for IgorP
 
 For further detail on individual functions, consult the [G3F API](https://github.com/dap-biospec/G3F/blob/master/Docs/G3F_API.md)
 
-### **LOADING AND OPENING G3F IN IGOR PRO:**
+### **Installation**
 
 The G3F package consists of several files:
 
@@ -28,7 +28,7 @@ Download all files from [G3F folder](https://github.com/dap-biospec/G3F/blob/mas
 From the top menu IgorPro:<br/>
 File -> Open File -> Procedure…
 
-Navigate to and select the G3F\_Main.ipf file to load. Complile G3F\_Main.ipf - this will load all dependnet procedure file slisted above.
+Navigate to and select the G3F\_Main.ipf file to load. Complile G3F\_Main.ipf - this will load all dependnet procedure files listed above.
 
 **Opening main G3F control panel:**
 
@@ -348,11 +348,15 @@ The results of this fit are stored in data structures for all parameters (global
 
 ### Demo Experiment
 
-To demonstrate the efficacy of G3F, an example NPSV experiment is provided with the G3F package.  [Demo Experiment](https://github.com/dap-biospec/G3F/tree/master/Demo)This demo models the extraction of the redox profile of myoglobin with different spectra taken at different applied potentials. In addition to different applied potentials, different compositions of 2 different electrochemical mediators (methylene green, thioninie acetate) are present in this data. In order to analyze the redox profiles of myoglobin alone, the data must be deconvoluted.
+To demonstrate the efficacy of G3F, an example NPSV experiment is provided with the G3F package.  [Demo Experiment](https://github.com/dap-biospec/G3F/tree/master/Demo) This demo models the extraction of the redox profile of myoglobin with different spectra taken at different applied potentials. In addition to different applied potentials, different compositions of 2 different electrochemical mediators (methylene green, thioninie acetate) are present in this data. In order to analyze the redox profiles of myoglobin alone, the data must be deconvoluted. This analysis is demonstrated in three separate ways in order to familiarize the user with the flexibility of the G3F package.
+
+- [Example #1: Direct Fitting](#direct-fitting)
+- [Example #2: Process Fitting](#process-fitting)
+- [Example #3: Calculating Fits Using Local Variables](#calculating-fits-using-local-variables)
 
 **Installing and Opening:**
 
-Download G3F as described in the Installation section.
+Download G3F as described in the [Installation section](#installation).
 
 Download folder and save as a sibling folder of the G3F folder.
 
@@ -367,6 +371,10 @@ Open the main control panel, and type in the values for the global variables:
 ![alt text](https://github.com/dap-biospec/G3F/blob/master/Demo/DemoPictures/Pic_1_Demo.png)
 
 The control panel should be set up for package testing.
+
+### **Direct Fitting**
+
+In this example, the data will be fitted directly using a fitting function and global variables for known parameters of analytes (ie: standard reduction potentials, number of electrons transferred). Individual spectra corresponding to each analyte and the Nernstian profile of myoglobin are calculated using local variables.
 
 **Loading Raw Data and Fitting Function**
 
@@ -442,4 +450,199 @@ By formatting the raw data as dots and keeping the fit solid, a graph should be 
 
 From these traces, the redox potential of myoglobin can be determined. The multidimensional data has been successfully deconvoluted.
 
- 
+### **Process Fitting**
+
[…]
