### 2019-07-30 bbc0a85
Update G3F_API.md
