### 2019-07-30 757d2e7
Update readme_performance_test.md
