### 2019-07-31 31d764f
Update readme.md
Updated readme to include essential tutorial steps (ie: local parameter guesses)
