### 2019-07-31 0a21e7d
Update readme.md
Added MediatorSpectra picture for tutorial.
