### 2019-07-31 4abaeff
Add files via upload
Updated manual to include tutorial fixes (local parameter guesses, new demo pictures for tutorial).
