+++ Docs/G3F_Manual.md
@@ -1,4 +1,4 @@
-## G3F MANUAL
+﻿## G3F MANUAL
 
 ### Introduction
 
@@ -346,201 +346,6 @@ When all necessary parameters have been set on the control panel, the fit may be
  
 The results of this fit are stored in data structures for all parameters (global variables, local variables, and fitted data) for easy extraction, plotting and analysis by the user.
 
-### Demo Experiment
-
-To demonstrate the efficacy of G3F, an example NPSV experiment is provided with the G3F package.  [Demo Experiment](https://github.com/dap-biospec/G3F/tree/master/Demo) This demo models the extraction of the redox profile of myoglobin with different spectra taken at different applied potentials. In addition to different applied potentials, different compositions of 2 different electrochemical mediators (methylene green, thioninie acetate) are present in this data. In order to analyze the redox profiles of myoglobin alone, the data must be deconvoluted. This analysis is demonstrated in three separate ways in order to familiarize the user with the flexibility of the G3F package.
-
-- [Example #1: Direct Fitting](#direct-fitting)
-- [Example #2: Process Fitting](#process-fitting)
-- [Example #3: Calculating Fits Using Local Variables](#calculating-fits-using-local-variables)
-
-**Installing and Opening:**
-
-Download G3F as described in the [Installation section](#installation).
-
-Download folder and save as a sibling folder of the G3F folder.
-
-Open Demo_G3F.pxp
-
-Experiment should load G3F code automatically. If G3F_main.ipf cannot be found, IgorPro will present a dialog that will allow the user to navigate to the G3F folder.
-
-Demo experiment will also load the SpecEchem_4Spec_ForOxd_2.ipf procedure from the Demo folder with the example of user-supplied fitting function. 
-
-Open the main control panel, and type in the values for the global variables:
-
-![alt text](https://github.com/dap-biospec/G3F/blob/master/Demo/DemoPictures/Pic_1_Demo.png)
-
-The control panel should be set up for package testing.
-
-### **Direct Fitting**
-
-In this example, the data will be fitted directly using a fitting function and global variables for known parameters of analytes (ie: standard reduction potentials, number of electrons transferred). Individual spectra corresponding to each analyte and the Nernstian profile of myoglobin are calculated using local variables.
-
-**Loading Raw Data and Fitting Function**
-
-Go to the **Dataset** section on the control panel. Use the drop-down menu to select &quot;Oxidation.&quot; This loads the raw data matrix of oxidation data.
-
-Go to the **Fit to function** drop-down menu. Select &quot;SpecEChem_4Spec_ForOxd_2D&quot;
-
-**Running the Fit:**
-
-The global parameters displayed on the control panel correspond to the redox potentials and electron transfer coefficients of each mediator.
-
-![alt text](https://github.com/dap-biospec/G3F/blob/master/Demo/DemoPictures/Pic_2_Demo.png)
-
-As there are 2 mediators and myoglobin present, the data must be deconvoluted to isolate the spectrum and Nernstian profile of myoglobin. To do this, four local parameters are set (methylene green has two electron transfer steps), each corresponding to a spectrum specific to the analyte.
-
-![alt text](https://github.com/dap-biospec/G3F/blob/master/Demo/DemoPictures/Pic_3_Demo.png)
-
-In this experiment, all baselines are set to zero (Col(Z) Locals). These are not used in this simulation.
-
-Next, holds must be set on known parameters. The electron transfer coefficients correspond to the number of electrons transferred per mediator and are known.
-
-Go to the **Matrix Constraints** button.
-
[…]
