### 2019-07-31 c4516d5
Update of demo and performance testing
- Demo experiment no includes three separate modes of fitting. Documentation has been updated accordingly.
- Included separate performance testing experiment and documentation on verifying fitting results under pre-configured conditions.
