### 2019-07-31 f89bf83
Update of demo and performance testing
- Demo experiment no includes three separate modes of fitting. Documentation has been updated accordingly.
- Included separate performance testing experiment and documentation on verifying results under pre-configured conditions.
