+++ Docs/G3F_Manual.md
@@ -326,25 +326,213 @@ The column parameters wave and row parameters wave will contain calculated local
 
 This procedure is a simple Lorentzian function for several isotopes with local variables ([iCp][1, 2, ..N]) describing the amplitude. The line: `wY[i]  = I_a * wCP2D[iCP][0] * wGP[wCClb[i]+4] * lorentzian(wl, n162_a, Gw_a)` describes the intensity ( `I_a`) modified by a local 2D column parameter ( `wCP2D[iCP][0]`). The following statement (`wGP[wCClb[i]+4]`) describes the column calibration parameter, followed by the lorenztian model with global parameters.
 
-**To load a fitting function:**
+### Demo Experiment
 
-On the **Method** section on the Control Panel, the user may select a pre-compiled fitting function on the drop-down menu **Fit to function** as shown below:
+To demonstrate the efficacy of G3F, an example NPSV experiment is provided with the G3F package.  [Demo Experiment](https://github.com/dap-biospec/G3F/tree/master/Demo) This demo models the extraction of the redox profile of myoglobin with different spectra taken at different applied potentials. In addition to different applied potentials, different compositions of 2 different electrochemical mediators (methylene green, thioninie acetate) are present in this data. In order to analyze the redox profiles of myoglobin alone, the data must be deconvoluted. This analysis is demonstrated in three separate ways in order to familiarize the user with the flexibility of the G3F package in data analysis.
 
-![alt text](https://github.com/dap-biospec/G3F/blob/master/Docs/G3FManualPictures/LoadFunction.png)
+- [Example #1: Direct Fitting](#direct-fitting)
+- [Example #2: Process Fitting](#process-fitting)
+- [Example #3: Calculating Fits Using Local Variables](#calculating-fits-using-local-variables)
 
-The function selected in the **Fit to function** dialog will be used to fit experimental data.
+**Installing and Opening:**
 
-### Loading and Fitting Data
+Download G3F as described in the [Installation section](https://github.com/dap-biospec/G3F/blob/master/Docs/G3F_Manual.md#installation).
 
-Three-dimensional datasets may be loaded as a 3D matrix into IgorPro and selected using the **Dataset** dropdown menu as shown below:
+Download folder and save as a sibling folder of the G3F folder.
 
-![alt text](https://github.com/dap-biospec/G3F/blob/master/Docs/G3FManualPictures/LoadData_1.png) 
+Open Demo_G3F.pxp
 
-When all necessary parameters have been set on the control panel, the fit may be tested using the **Simulate** button on the **Global fit** tab on the main control panel interface. If the results are satisfactory, the fit can be performed by clicking the **Do fit now!** button.
+Experiment should load G3F code automatically. If G3F_main.ipf cannot be found, IgorPro will present a dialog that will allow the user to navigate to the G3F folder.
 
-![alt text](https://github.com/dap-biospec/G3F/blob/master/Docs/G3FManualPictures/DoFit_Button.png)
- 
-The results of this fit are stored in data structures for all parameters (global variables, local variables, and fitted data) for easy extraction, plotting and analysis by the user.
+Demo experiment will also load the SpecEchem_4Spec_ForOxd_2.ipf procedure from the Demo folder with the example of user-supplied fitting function. 
+
+Open the main control panel, and type in the values for the global variables:
+
+![alt text](https://github.com/dap-biospec/G3F/blob/master/Demo/DemoPictures/Pic_1_Demo.png)
+
+The control panel is now set up for package testing.
+
+### **Direct Fitting**
+
[…]
