### 2019-07-31 78149a2
Update of demo and performance testing
-Demo now includes three separate modes of fitting. Documentation has been updated accordingly.
-Included separate performance testing experiment and documentation on verifying fitting results under pre-configured conditions.
