+++ Demo/readme.md
@@ -1,6 +1,6 @@
 ### Demo Experiment
 
-To demonstrate the efficacy of G3F, an example NPSV experiment is provided with the G3F package.  [Demo Experiment](https://github.com/dap-biospec/G3F/tree/master/Demo) This demo models the extraction of the redox profile of myoglobin with different spectra taken at different applied potentials. In addition to different applied potentials, different compositions of 2 different electrochemical mediators (methylene green, thionine acetate) are present in this data. In order to analyze the redox profiles of myoglobin alone, the data must be deconvoluted. This analysis is demonstrated in three separate ways in order to familiarize the user with the flexibility of the G3F package in data analysis.
+To demonstrate the efficacy of G3F, an example NPSV experiment is provided with the G3F package.  [Demo Experiment](https://github.com/dap-biospec/G3F/tree/master/Demo) This demo models the extraction of the redox profile of myoglobin with different spectra taken at different applied potentials. In addition to different applied potentials, different compositions of 2 different electrochemical mediators (methylene green, thioninie acetate) are present in this data. In order to analyze the redox profiles of myoglobin alone, the data must be deconvoluted. This analysis is demonstrated in three separate ways in order to familiarize the user with the flexibility of the G3F package in data analysis.
 
 - [Example #1: Direct Fitting](#direct-fitting)
 - [Example #2: Process Fitting](#process-fitting)
@@ -22,7 +22,7 @@ Open the main control panel, and type in the values for the global variables:
 
 ![alt text](https://github.com/dap-biospec/G3F/blob/master/Demo/DemoPictures/Pic_1_Demo.png)
 
-The control panel is now set up for package testing.
+The control panel should be set up for package testing.
 
 ### **Direct Fitting**
 
@@ -56,11 +56,13 @@ Hold all variables (K0 through K5; K8 through K10) by checking the **Hold** fiel
 
 Since the spectra of the known mediators and the anlyte will be calculated using row locals, it is necessary to provide guesses for these variables. To do this, open the data browser and select MediatorSpectra. Highlight the spectra as shown and copy it.
 
-![alt text](https://github.com/dap-biospec/G3F/blob/master/Demo/DemoPictures/Pic_18_Demo.PNG)
+![alt text](https://github.com/dap-biospec/G3F/blob/master/Demo/DemoPictures/Pic_18_Demo.png)
 
-Paste this spectra into the RowLoc field.
+Paste this spectra into the Oxidation_RowLoc field.
 
 ![alt text](https://github.com/dap-biospec/G3F/blob/master/Demo/DemoPictures/Pic_6_Demo.png)
+
+Go to the **Options** tab. Check the box called **Epsilon**. This allows for the application of epsilon to matrix constraints. While not required, 2D or 3D global fits will **generally fail if epsilon is ignored**.
 
 To perform the fit, click **Do Fit Now!** The program will run for a few minutes and generate the fitted data.
 
@@ -80,7 +82,7 @@ This fitted data set corresponds to each individual spectrum of each mediator an
 
 Windows ->New Graph…
 
-Select Oxidation\_RowLocal in the left field, and Oxidation\_ROW\_FIT in the right field. Add the four mediator spectra as shown below.
+Select Oxidation\_RowLoc in the left field, and Oxidation\_ROW\_FIT in the right field. Add the four mediator spectra as shown below.
 
 ![alt text](https://github.com/dap-biospec/G3F/blob/master/Demo/DemoPictures/Pic_7_Demo.png)
 
@@ -144,11 +146,13 @@ Hold all variables (K0 through K5; K8 through K10) by checking the **Hold** fiel
 
[…]
