### 2019-07-31 #1568 @dap-biospec
@tcausgrove, @kyleniemeyer - Documentation has been updated to keep repo transparent. Additional examples are provided in the demo experiment to better illustrate capabilities in three scenarios. A separate performance testing experiment is provided along with the description of testing process and the expected results.
