### 2019-08-05 #1568 @FaustinCarter
In theory, yes. In practice, it depends on whether I can get IgorPro 8 to give me another trial period. I have an older license, but this code does not appear to be backwards compatible. Give me a couple days and I'll have a definite. If you find a different reviewer in the meantime, that is also fine.
