### 2019-08-05 #1568 @dap-biospec
@FaustinCarter - I would greatly appreciate it. Please let me know if we can help you with IgorPro 8 in any way. For example, we can set up a remote desktop to one of our computers with IgorPro running site license. I can also reach out to Wavemetrics if you would like.
