### 2019-08-05 #1568 @FaustinCarter
Either one of those sound like reasonable options to me. If you can make
one of them work I'm on board.

On Mon, Aug 5, 2019 at 12:09 PM dap-biospec <notifications@github.com>
wrote:

> @FaustinCarter <https://github.com/FaustinCarter> - I would greatly
> appreciate it. Please let me know if we can help you with IgorPro 8 in any
> way. For example, we can set up a remote desktop to one of our computers
> with IgorPro running site license. I can also reach out to Wavemetrics if
> you would like.
>
