### 2019-08-06 #1568 @dap-biospec
@FaustinCarter  I am asking our IT staff to set up a limited time VPN access for you (I have control of my group our workstations). What email address would you like to use? It would be better if it is more official than GMail (not for me, for administration). If you do not want to post it here, email me directly at dapro@chemistry.msu.edu.

Thanks for understanding!
