### 2019-08-06 #1629 @whedon
[REVIEW]: G3F: Global, Multidimensional Spectral Regression Analysis
**Submitting author:** @dap-biospec (<a href="http://orcid.org/0000-0003-4625-0323">Denis Proshlyakov</a>)
**Repository:** <a href="https://github.com/dap-biospec/G3F" target ="_blank">https://github.com/dap-biospec/G3F</a>
**Version:** v1.0a4

- [ ] **Version:** Does the release version given match the GitHub release (v1.0a4)?

- [ ] **Version:** Does the release version given match the GitHub release (v1.0a4)?
