### 2019-08-10 9d7d4b9
Updated DOIs
All citations include DOI
Added missing ChemRxiv reference
