+++ paper.bib
@@ -4,6 +4,7 @@
 	Month = jan,
 	Title = {{Insight into the mechanism of an iron dioxygenase by resolution of steps following the Fe(IV)=O species}},
 	Year = 2010,
+	doi = {10.1073/pnas.0911565107},
 	adsurl = {https://www.pnas.org/content/107/9/3982/},
 }
 
@@ -13,34 +14,39 @@
 	Month = feb,
 	Title = {{Structure of the key species in the enzymatic oxidation of methane to methanol}},
 	Year = 2015,
+	doi = {10.1038/nature14160},	
 	adsurl = {https://www.nature.com/articles/nature14160#main},
 }
 
 @article{John:2019a,
 	Author = {{John}, C.~W. and {Proshlyakov}, D.~A.},
 	Journal = {Analytical Chemistry},
-	Month = july,
 	Title = {{FTIR Spectro-Voltammetry and Quantitative Modeling of Analytes in Kinetically Constrained Redox Mixtures}},
 	Year = 2019,
+	doi = {10.1021/acs.analchem.9b00859},
 	adsurl = {https://pubs.acs.org/doi/10.1021/acs.analchem.9b00859}
 }
 
 @article{John:2019b,
-	Adsnote = {Placeholder for paper under intial review in The Journal of Physical Chemistry B},
+	Adsnote = {submitted to The Journal of Physical Chemistry B},
 	Author = {{John}, C.~W., and {Swain}, G.~M. and {Hausinger}, R.~P. and {Proshlyakov}, D.A.},
-	Journal = {The Journal of Physical Chemistry B},
+	Journal = {ChemRxiv e-prints},
 	Title = {{Strongly coupled redox-linked conformational switching at the active site of the non-heme iron-dependent dioxygenase, TauD}},
-	Year = 2019
+	Year = 2019,
+	doi = {10.26434/chemrxiv.9461462}
 }
+
 
 @book{Anderson:2011,
 	Author = {{Anderson}, D. and {Kurtz}, T.},
 	Booktitle = {Design and Analysis of Biomolecular Circuits: Engineering Approaches to Systems and Synthetic Biology, by H. Koeppl and H. Densmore and G. Setti and M. di Bernardo.~ISBN 978-1-4419-6766-4.~Published by Springer, 2011.},
 	Publisher = {Springer},
 	Title = {{Continuous Time Markov Chain Models for Chemical Reaction Networks}},	
-	Year = 2011
+	Year = 2011,
+	adsurl = {https://doi.org/10.1007/978-1-4419-6766-4_1},
+	doi = {10.1007/978-1-4419-6766-4_1}
 }
-	
+
 @article{Zhang:2018,
 	Author = {{Zhang}, C. and {Shen}, Z. and {Wei}, W. and {Zhao}, J. and {Zhang}, Z., and {You}, X.},
 	Eprint = {1802.05170},
@@ -49,5 +55,5 @@
 	Title = {{Molecular Computing for Markov Chains}},
 	Year = 2018,
 	adsurl = {https://arxiv.org/abs/1802.05170},
+	doi = {10.1007/s11047-019-09736-8}
 }	
-
