### 2019-08-11 #1629 @tcausgrove
I'm working through the demo, starting to understand the software.  I'm having two issues, which may be related.  One, when I open the demo, Igor looks for a "submission_02_works:Demo" folder, which I can't find.  (The folder levels in general seem off, extra ::: in front.)  Also, for the Process Fitting part of the demo, I can't find the Process_SpecEChem_4Spec_ForOxd_2D function.  Am I missing something?
