### 2019-08-11 #1629 @dap-biospec
@tcausgrove I updated Demo_G3F.pxp and Performance_Test.pxp after your earlier message. The  offending path was deleted. I recommend downloading  complete repo as ZIP file to ensure that Igor experiments are current. A characteristic change is checked "Epsilon" checkbox in the main panel.  Another check is Misc->"Path Status": QuantAnalysis path should be "[path_to_parent]\G3F-master\G3F\".

The process fitting function Process_SpecEChem_4Spec_ForOxd_2D (and others required for methods 2 and 3) in now in the updated  SpecEchem_4Spec_ForOxd_2.ipf procedure file.
