### 2019-08-12 #1629 @dap-biospec
@tcausgrove, @FaustinCarter  - we updated the demo to exclude unnecessary path (error on load) and to use the correct demo procedure file. Thank you @tcausgrove for bringing it up.  Although we tested full download from GitHub before the review, some procedures were still loaded from our network locations instead of relative location in the repo mirror.  We identified the cause only after disabling the network on the testing computer.  Sorry about that!

Since .gitignore includes *.pxp entry, please make sure that you have the latest demo and testing experiments. Re-downloading repository as ZIP should do the trick.
