### 2019-08-13 #1629 @tcausgrove
OK, the demo runs all the way through for me now.  Thanks @dap-biospec for the new version.  It has taken me a while to figure out how things fit together - there are quite a few moving parts to this software.  The learning curve might be a bit of a barrier to adoption.

There are a few things that could probably be addressed more thoroughly in the manual.  For example, the items in the Analysis->Global 3D Spectral Regression menu are barely mentioned, but seem very useful.

Some things that I think would be good for future releases:

1. Expanded manual
2. A demo using layers
3. Expanded description of the autocycle feature
4. Clean up demo - not sure why all of the Pulse... waves are there, how the Reduction data set fits in, or why it uses a 5th order polynomial for the background when all the coefficients are zero
5. Example of post-processing
6. Example of using reference data

This is a really extensive system for fitting multi-dimensional data.  The software seems to be more or less in finished form, but the documentation could be improved as an ongoing effort.  I am ready to recommend acceptance.
