### 2019-08-13 #1629 @dap-biospec
Thank you, @tcausgrove. We certainly will continue to develop tutorial moving forward. We are preparing a publication where layers are instrumental for describing isotopic compositions of  time-resolved vibrational spectra and we think that this would be a great context  to showcase this.

Much of the demo is a simple example taken out of context of our two papers cited herein. We will check for and purge unnecessary data and add a note on the context to the tutorial.

Again, thank you for a speedy review!

cc:  @FaustinCarter
