### 2019-08-13 06f4633
clarification of optional data and methods
