+++ Docs/G3F_Manual.md
@@ -64,9 +64,11 @@ The main control panel of the G3F macro is split into several different sections
 
 **Reference Data:**
 
-This section is composed of additional data needed for fitting calculations. This is an optional feature for fitting.
+This section is composed of additional data that may be supplied for fitting calculations. This can be any arbitrary data and its interpretation is entirely up to the user and the specfic needs of the model. It is important to note that including any reference data changes the process (if used), fitting, and post-processing template function. For example, in direct calculation method G3F_Direct_1D_TPL template function will be changed to G3F_Direct_Ep_1D_TPL when reference data are included.
 
-- Global reference data: optional numeric wave passed to process, fitting, and post-processing functions
+- Global reference data: optional numeric wave passed to process, fitting, and post-processing functions. Selected wave can be of any dimension, including zero. An example of this type of data would be tables of thermodynamic constants.
+
+- Column and row reference data: their row dimensions must match the row/column dimensions of the fitted data set. Examples of this kind of data are: reference spectra, and redox response profiles. A Global reference data wave (possibly of zero size) must be supplied to enable local reference data option.
 
 **Method:**
 
@@ -80,7 +82,7 @@ Once the process is generated, it is passed to the fitting function that perform
 - Last Process:
   - Fit to function: This is the fitting function that is being used for 3DF fit. This function defines the global variables and local variables used in the fit.
 
-- Post-processing: Post-processing functions can perform additional processing after all fittings have been performed. These functions must be supplied by the user.
+- Post-processing: this optional, user-supplied function can perform any additional processing of the fitted results. Examples include combining fitting results, integration of spectral data, calculation of ratios, etc. This function is executed once after the fitting or simulation is complete without an error.
 
 **Dataset:**
 
@@ -193,7 +195,7 @@ This section allows the user to override specific holds for each dimension, glob
 
 **Autocycle Hold:**
 
-This section allows the user to hold variables after a predetermined number of automatic cycles.
+For particularly complex fitting models, the global regression may not converge unless certain variables are alternately held and unheld during fitting. This also can prevent NAN and singular matrix errors. This section allows the user to alternate fitting of only individual subsets of variables (global, column, row etc) while all others are held. This feature is especially helpful if intial guesses are far from optimal. Generally, provided with a sufficient number of iterations, auto-cycling will converge to the same result with higher robustness at the expense of increased computational time. 
 
 - cycles: Sets the number of automatic cycles to be performed using one of combinations detailed below.
 - Global – Row – Col: Alternate analysis of one group at a time between Global, Row local and Column local variables.
@@ -328,11 +330,13 @@ This procedure is a simple Lorentzian function for several isotopes with local v
 
 ### Demo Experiment
 
-To demonstrate the efficacy of G3F, an example NPSV experiment is provided with the G3F package.  [Demo Experiment](https://github.com/dap-biospec/G3F/tree/master/Demo) This demo models the extraction of the redox profile of myoglobin with different spectra taken at different applied potentials. In addition to different applied potentials, different compositions of 2 different electrochemical mediators (methylene green, thioninie acetate) are present in this data. In order to analyze the redox profiles of myoglobin only, the data must be deconvoluted. This analysis is demonstrated in three separate ways in order to familiarize the user with the flexibility of the G3F package in data analysis.
[…]
