### 2019-08-13 939aea8
clarifications of the demo models
