### 2019-08-13 #1629 @dap-biospec
@tcausgrove, @FaustinCarter Non-essential data were purged from the demo experiment.
Explanations were added to the demo and manual documents to clarify the use of reference data, post-processing functions, and auto-cycling of parameter holds overrides.
