+++ Demo/readme.md
@@ -5,8 +5,10 @@ To demonstrate the efficacy of G3F, an example normal-pulse staircase voltammetr
 - [Example #1: Direct Fitting](#direct-fitting)
 - [Example #2: Process Fitting](#process-fitting)
 - [Example #3: Calculating Fits Using Local Variables](#calculating-fits-using-local-variables)
+- [Example #4: Calculating Fits Using Reference Data](#calculating-fits-using-reference-data)
+- [Example #5: Calculating Fits Using Post Processing Methods](#calculating-fits-using-post-processing-methods)
 
-The dataset provided here consists of global and local variables, as detailed in the manual. Each experimental spectrum is contributed by multiple species. In addition, it includes experimental error, which is described by a polynomial baseline.  In the examples 1 and 2, terms of the polynomial are fitted as the column local variables. Please note that the fifth order polynomial, shown in this demo, is not necessary for all datasets; this can be modified or substituted to match the needs of the particular quantitative model. In this demo, vibrational spectra are described by the row local variables (row locals), with each column of the row local wave corresponding to a vibrational spectrum of one species. Known parameters (i.e. the number of electrons and the redox potentials of the mediators) are represented by global variables.
+The dataset provided here consists of global and local variables, as detailed in the manual. Each experimental spectrum is contributed by multiple species. In addition, it includes experimental error, which is described by a polynomial baseline.  In the examples 1 and 2, terms of the polynomial are fitted as the column local variables. Please note that the fifth order polynomial, shown in this demo, is not necessary for all datasets; this can be modified or substituted to match the needs of the particular quantitative model. In this demo, vibrational spectra are described by the row local variables (row locals), with each column of the row local wave corresponding to a vibrational spectrum of one species. Known parameters (i.e. the number of electrons and the redox potentials of the mediators) are represented by global variables. In Example 4, known parameters (i.e. i.e. the number of electrons and the redox potentials of species) are supplied using reference data in order to generate thermodynamic properties of all species in the final fit.
 
 **Installing and Opening:**
 
@@ -331,4 +333,226 @@ By formatting the raw data as dots and keeping the fit solid, a graph should be 
 ![alt text](https://github.com/dap-biospec/G3F/blob/master/Demo/DemoPictures/Pic_11_Demo.png)
 
 From these traces, the redox potential of myoglobin can be determined. The multidimensional data has been successfully deconvoluted as in Example 1 and Example 2.
+
+### **Calculating Fits Using Reference Data**
+
+In this example, the data will be fitted using a reference global and row data (ExtraGlobal and ExtraRow waves respectively). The spectrum and population of myoglobin (Species D) are fitted in this example, with mediator (methylene green, thioninie acetate) spectra supplied using the ExtraRow wave. This approach is intended to allow the user to implement known references in quantitative models.
+
+**Loading Raw Data and Fitting Function**
+
[…]
