### 2019-08-20 9fcc041
Extra parameters and post-processing examples
Demo adds extra parameters and post-processing examples. Tutorial updated accordingly.
