### 2019-08-20 dd32602
Extra parameters and post-processing examples
Demo adds extra parameters and post-processing examples. Manual updated accordingly.
