### 2019-08-20 7a9b048
Extra parameters and post-processing examples
Demo adds extra parameters and post-processing examples. Tutorial updated accordingly.
