### 2019-08-20 284b50b
Extra parameters and post-processing example
Demo adds extra parameters and post-processing example. Manual updated accordingly,
