+++ Docs/G3F_Manual.md
@@ -1,333 +1,3 @@
-﻿## G3F MANUAL
-
-### Introduction
-
-The G3F (Global 3-dimensional Fit) is a simulation and fitting package for IgorPro. It allows to perform global spectral regression using flexible combinations of global parameters and dimensionally local (spectral) parameters on up to 3-dimensional matrix datasets.
-
-**How to use this manual:** This manual summarizes the capabilities of the G3F software IgorPro macro. The functions of all features are briefly summarized and documented for easy reference. Basic knowledge of Igor Pro operation and data structures within Igor Pro is assumed. **For quick help, hover over functions on the control panel for simple tooltips.**
-
-For further detail on individual functions, consult the [G3F API](https://github.com/dap-biospec/G3F/blob/master/Docs/G3F_API.md)
-
-### **Installation**
-
-The G3F package consists of several files:
-
-- G3F\_Main.ipf
-- G3F\_Auxillary.ipf
-- G3F\_Desktop.ipf
-- G3F\_DoTheFit.ipf
-- G3F\_Feedback.ipf
-- G3F\_FitFlow.ipf
-- G3F\_Guesses.ipf
-- G3F\_List&amp;Set.ipf
-- G3F\_Setup.ipf
-- G3F\_Struct.ipf
-
-Download all files from [G3F folder](https://github.com/dap-biospec/G3F/blob/master/G3F/) and save in one local folder.
-
-From the top menu IgorPro:<br/>
-File -> Open File -> Procedure…
-
-Navigate to and select the G3F\_Main.ipf file to load. Complile G3F\_Main.ipf - this will load all dependnet procedure files listed above.
-
-**Opening main G3F control panel:**
-
-From the top menu IgorPro:<br/>
-Analysis -> Global 3D Spectral Regression -> Control Panel
-
-
-### Data Structure
-
-The overall data structure of the datasets handled by G3F is shown below:
-
-![alt text](https://github.com/dap-biospec/G3F/blob/master/Docs/G3FManualPictures/3D_DataStructure_B.JPG)
-
-Data sets consist of &quot;layers&quot; of matrices with &quot;row&quot;, &quot;column&quot; and &quot;layer&quot; calibrations. The 2D layers of this data structure are shown conceptually below:
-
-![alt text](https://github.com/dap-biospec/G3F/blob/master/Docs/G3FManualPictures/DataStructure_Simple.JPG)
-
-The dimensionality of this data structure is further expanded with Layer-Column (LCP2D) and Layer-Row (LRP2D) dimensions, as shown below:
-
-![alt text](https://github.com/dap-biospec/G3F/blob/master/Docs/G3FManualPictures/DataStructure_2.png)
- 
- An alternative depiction of this data structure with emphasis on the roles of global parameters and local parameters is shown below:
-
- ![alt text](https://github.com/dap-biospec/G3F/blob/master/Docs/G3FManualPictures/DataStructure_3.jpg)
- 
-Row, column, and layer locals are calculated parameters, whereas global parameters are parameters which remain consistent throughout the data set.
-
-### Summary of G3F User Interface
-
-![alt text](https://github.com/dap-biospec/G3F/blob/master/Docs/G3FManualPictures/ControlPanel.png)
-
-The main control panel of the G3F macro is split into several different sections, each of which are detailed below:
-
-**Reference Data:**
-
-This section is composed of additional data that may be supplied for fitting calculations. This can be any arbitrary data and its interpretation is entirely up to the user and the specfic needs of the model. It is important to note that including any reference data changes the process (if used), fitting, and post-processing template function. For example, in direct calculation method G3F_Direct_1D_TPL template function will be changed to G3F_Direct_Ep_1D_TPL when reference data are included.
-
[…]
