### 2019-08-21 92262c7
Extra parameters and post-processing examples
Demo adds extra parameters and post-processing examples. Tutorial updated accordingly.
