### 2019-08-21 894e2a8
Update G3F_Manual.md
