### 2019-08-21 0cc4e0a
Extra parameters and post-processing examples.
Demo adds extra parameters and post-processing examples. Manual updated accordingly.
