+++ Docs/G3F_Manual.md
@@ -334,15 +334,15 @@ This procedure is a simple Lorentzian function for several isotopes with local v
 
 ### Demo Experiment
 
-To demonstrate the efficacy of G3F, an example normal-pulse staircase voltammetry (NPSV) experiment determining the redox potential of myoglobin is provided with the G3F package.  [Demo Experiment](https://github.com/dap-biospec/G3F/tree/master/Demo) The redox potential is key to understanding the behavior of molecules, and in this example a protein. As the active site is typically buried deep within the protein, electrochemical experiments are often conducted using mediators. For further information on this type of analysis, see (https://pubs.acs.org/doi/10.1021/acs.analchem.9b00859). This demo models the analysis of the vibrational (FTIR) redox response of myoglobin from spectra taken at different applied potentials. Myoglobin and each mediator (methylene green, thioninie acetate)  present in the medium possess their own vibrational spectra and redox potentials. In order to interpret the redox and vibrational signatures of myoglobin only, the data must be deconvoluted. Three separate methods of analysis are demonstrated in order to familiarize the user with the flexibility of the G3F package in data analysis.
+To demonstrate the efficacy of G3F, an example normal-pulse staircase voltammetry (NPSV) experiment determining the redox potential of myoglobin is provided with the G3F package.  [Demo Experiment](https://github.com/dap-biospec/G3F/tree/master/Demo) The redox potential is key to understanding the behavior of molecules, and in this example a redox-active protein. As the active site is typically buried deep within the protein, electrochemical experiments are often conducted in the presence of mediators. For further information on this example, see (https://pubs.acs.org/doi/10.1021/acs.analchem.9b00859). This demo exemplifies the analysis of the vibrational (FTIR) redox response of myoglobin from spectra taken at different applied potentials. Myoglobin and each mediator (methylene green, thioninie acetate),  present in the medium, possess their own vibrational spectra and redox potentials. In order to interpret the redox and vibrational signatures of myoglobin only, the data must be deconvoluted. Three separate methods of analysis are demonstrated here in order to familiarize the user with the flexibility of the G3F data analysis.
 
 - [Example #1: Direct Fitting](#direct-fitting)
-
 - [Example #2: Process Fitting](#process-fitting)
-
 - [Example #3: Calculating Fits Using Local Variables](#calculating-fits-using-local-variables)
+- [Example #4: Calculating Fits Using Reference Data](#calculating-fits-using-reference-data)
+- [Example #5: Calculating Fits Using Post Processing Methods](#calculating-fits-using-post-processing-methods)
 
-The dataset provided consists of global and local variables, as detailed in the manual. Each experimental spectrum is contributed by multiple species. In addition, it includes experimental error, which is described by a polynomial baseline.  In the examples 1 and 2, the terms of a polynomial are fitted as the column local variables. Please note that the fifth order polynomial shown in this demo is not necessary for all datasets; this can be modified or substituted to match the needs of the particular quantitative model. In this demo vibrational spectra are described by the row local variables (row locals), with each column of the row local wave corresponding to the vibrational spectra of one species. Known parameters (i.e. the number of electrons and the redox potentials of the mediators) are represented by global variables.
[…]
