### 2019-08-21 c77fe78
Extra parameters and post-processing examples
Demo adds extra parameters and post-processing examples. Tutorial updated accordingly.
