### 2019-08-21 dd2f1dd
Extra parameters and post-processing examples
Demo adds extra parameters and post-processing examples. Manual updated accordingly.
