+++ Docs/G3F_Manual.md
@@ -392,6 +392,8 @@ Go to the **Matrix Constraints** button.
 
 Hold all variables (K0 through K5; K8 through K10) by checking the **Hold** field. These correspond to the electron transfer coefficients and redox potentials of the mediators – all of which are known. The last two fields correspond to the electron transfer coefficient and redox potential of myoglobin. When finished, press **Done**.
 
+**Prior to continuing the demo, hit Simulate! This generates the needed fields for this and all following demos!**
+
 Since the spectra of the known mediators and the anlyte will be calculated using row locals, it is necessary to provide guesses for these variables. To do this, open the data browser and select MediatorSpectra. Highlight the spectra as shown and copy it.
 
 ![alt text](https://github.com/dap-biospec/G3F/blob/master/Demo/DemoPictures/Pic_18_Demo.PNG)
@@ -432,7 +434,11 @@ To extract the Nernstian profiles, go to the **Feedback positions** field on the
 
 ![alt text](https://github.com/dap-biospec/G3F/blob/master/Demo/DemoPictures/Pic_9_Demo.png)
 
-Click **Do Fit!** This will generate Nernstian curve waves for both experimental data and fits. To plot this data, go to:
+Click **Do Fit!** This will generate Nernstian curve waves for both experimental data and fits.
+
+**If this fit fails, set all fields in the Oxidation_ColLoc matrix to 0. This will likely solve the error**
+
+To plot this data, go to:
 
 Windows -> New Graph…
 
@@ -812,7 +818,11 @@ Since G3F calculates local variables by treating data waves as vectors, reasonab
 
 Click **No**.
 
-Open the **Data browser** and open the **Oxidation\_sim** wave from Example #2. If the **Keep** field was checked during Example 2, this wave will contain population calculations for analytes. If this wave is not present, re-run Example 2 with the **Keep** field in the **Method** tab checked.
+**FOR THIS EXAMPLE ONLY: An error message will display, showing a singular matrix error. Acknowledge the error and continue setup.**
+
+Open the **Data browser** and open the **Oxidation\_sim** wave from Example #2. If the **Keep** field was checked during Example 2, this wave will contain population calculations for analytes. If this wave is not present, re-run Example 2 with the **Keep** field in the **Method** tab checked. 
+
+Alternatively, the user can input data from the Reduction_PRef wave as initial guesses for the population profile. If this method is chosen, copy/paste the data into Column 06 as in Example 3 and skip to **Running the Fit** below.
 
 ![alt text](https://github.com/dap-biospec/G3F/blob/master/Demo/DemoPictures/Pic_16_Demo.PNG)
 
@@ -900,9 +910,7 @@ To extract the thermodynamic properties of myoglobin, go to the data browser, an
 
 ![alt text](https://github.com/dap-biospec/G3F/blob/master/Demo/DemoPictures/Pic_24_Demo.JPG)
 
-These values correspond to the redox potential, the number of electrons transferred, a constant necessary for the function of the Nernstenian fit, and the amplitude of the fit. These values are all generated using the Nernstenian post-processing function.
-
-
+These values correspond to the redox potential, the number of electrons transferred, a constant necessary for the function of the Nernstenian fit, and the amplitude of the fit. These values are all generated using the Nernstenian post-processing function. These values are also printed in the command line.
 
 
 
