+++ paper.bib
@@ -28,13 +28,14 @@
 }
 
 @article{John:2019b,
-	Adsnote = {submitted to The Journal of Physical Chemistry B},
 	Author = {{John}, C.~W., and {Swain}, G.~M. and {Hausinger}, R.~P. and {Proshlyakov}, D.A.},
-	Journal = {ChemRxiv e-prints},
+	Journal = {Journal of Physical Chemistry B},
 	Title = {{Strongly coupled redox-linked conformational switching at the active site of the non-heme iron-dependent dioxygenase, TauD}},
 	Year = 2019,
-	doi = {10.26434/chemrxiv.9461462}
+	doi = {10.1021/acs.jpcb.9b05866},
+	adsurl = {https://pubs.acs.org/doi/10.1021/acs.jpcb.9b05866}
 }
+
 
 
 @book{Anderson:2011,
