### 2019-08-24 #1629 @dap-biospec
Thank you @FaustinCarter. Point is taken. You are correct - this package transpired from a high-level chemical problems that could not be solved by conventional approaches.  We will continue to update documentation with more generic examples and additional scenarios.
