### 2019-08-25 eb2264d
correcting the paper
- multiple corrections have been made throughout the text
- description of the process fitting has been expanded and clarified.
