+++ paper.md
@@ -33,30 +33,31 @@ bibliography: paper.bib
 # Summary
 
 ## Rationale:
-Multi-dimensional non-linear global regression permits the investigation of quantitative relationships in complex datasets and to examine validity of proposed models. However, traditional multi-dimensional regression requires predictable variation of all parameters along every fitted dimension. This constraint may be difficult to satisfy if, for example:
+Multi-dimensional non-linear global regression permits the investigation of quantitative relationships in complex datasets and to examine validity of proposed models. However, traditional multi-dimensional regression requires predictable variation of all parameters along every fitted dimension. This constraint may be difficult to satisfy, for example:
 
-- When noise in any particular dimension exceed the signal of interest, such as with large polynomial variation of baselines due to thermal fluctuations, sample variability, or other of interference.  
-- When a common signal is too complex for rational description using a reduced set of variables, such as encountered with multiple bands in high-resolution spectra. The unknown spectra of one or more species involved in a predictable process may be better described by a large set of independent coefficients that vary between discrete sampled energies (frequency, wavelength, mass, charge etc.) than by a limited set of Gaussian or Lorentzian peaks. 
-- When observables cannot be accurately described using a trivial band shape or distribution, for example, when a normally distributed signal is broadened by spectral resolution with a rectangular profile. 
+- When noise in any particular dimension exceeds the signal of interest, such as with large variation of baselines due to thermal fluctuations, sample variability, or other of interference.  
+- When the common signal is too complex for rational description using a reduced set of variables, such as encountered with multiple bands in high-resolution spectra. The unknown spectra of one or more species, involved in a predictable process, may be better described by a large set of independent coefficients that vary between discrete sampled energies (frequency, wavelength, mass, charge etc.) than by a limited set of Gaussian or Lorentzian peaks. 
+- When observables cannot be accurately described using a trivial band shape or a distribution such as, for example, when a normally distributed signal is broadened by spectral resolution with a rectangular profile. 
 - When spectral overlap does not allow to achieve experimental resolution of several distinct signals, especially when their individual properties are not known.
-- When describing intensity an inhomogeneous kinetic process, a process with unknown kinetics, or when kinetics may be too complex be described with a reasonable number of phases.
+- When describing intensity in an inhomogeneous kinetic process, a process with an unknown kinetics, or when the kinetics may be too complex to be described by a reasonable number of phases.
 
 Common to these examples is the need to describe a multi-dimensional experimental dataset not only using globally invariable parameters (frequencies, temperature, rate constants), but with variable vectors of parameters that are applicable to one or more dimensions (local parameters).  
 
 ## Concept:
 
-The G3F package for IgorPro was initially developed to simultaneously analyze vibrational spectra of multiple isotopomers with overlapping modes in time-resolved Raman studies on enzyme TauD, where simple improvement of signal via time averaging was not possible. [@Grzyska:2010] Analysis of a frequency vs. time 2D dataset with uniform properties of predictable vibrational bandshapes (global) and unknown speciation plots (local to each time point)  allowed to resolve superimposed vibrations of two different species while improving resolution via signal sharing  between spectra. Also included were variable polynomial baseline, local to spectra. Similar approach was later used in analysis of Raman spectra of methane monooxygenase. [@Banerjee:2015] In more recent studies, it was expanded to obtain completely unknown difference infrared spectra of a redox transition with an unknown potential over a variable polynomial background.[@John:2019a; @John:2019b] In this case, analysis involved two orthogonal parametric vectors (frequency-dependent amplitudes vs. polynomial baselines). 
[…]
