### 2019-08-25 04f5a53
added general information
Added an introductory section on the general regression algorithms used by G3F, as well interpretation of the reported confidence values.
