### 2019-08-26 b0d72b9
update bibliography format
