+++ paper.bib
@@ -1,60 +1,68 @@
 @article{Grzyska:2010,
-	Author = {{Grzyska}, P. ~K. and {Appelman}, E. ~H and {Hausinger}, R. ~P and {Proshlyakov}, D. ~A.},
-	Journal = {PNAS},
-	Month = jan,
-	Title = {{Insight into the mechanism of an iron dioxygenase by resolution of steps following the Fe(IV)=O species}},
-	Year = 2010,
-	doi = {10.1073/pnas.0911565107},
-	adsurl = {https://www.pnas.org/content/107/9/3982/},
+ doi = {10.1073/pnas.0911565107},
+ year = 2010,
+ month = jan,
+ publisher = {United States National Academy of Sciences},
+ volume = {107},
+ number = {9},
+ pages = {3982--3987},
+ author = {Piotr K. Grzyska and Evan H. Appelman and Rovert P. Hausinger and Denis A. Proshlyakov},
+ title = {Insight into the mechanism of an iron dioxygenase by resolution of steps following the Fe(IV)=O species},
+ journal = {Proceedings of the National Academy of Sciences of the United States of America}
 }
 
 @article{Banerjee:2015,
-	Author = {{Banerjee}, R. and {Proshlyakov}, Y. and {Lipscomb}, J. ~D. and {Proshlyakov}, D. ~A.},
-	Journal = {Nature},
-	Month = feb,
-	Title = {{Structure of the key species in the enzymatic oxidation of methane to methanol}},
-	Year = 2015,
-	doi = {10.1038/nature14160},	
-	adsurl = {https://www.nature.com/articles/nature14160#main},
+ doi = {10.1038/nature14160},	
+ year = 2015,
+ month = feb,
+ publisher = {Springer Nature},
+ volume = {518},
+ pages = {431--434},
+ author = {Rahul Banerjee and Yegor Proshlyakov and John D. Lipscomb and Denis A. Proshlyakov},
+ title = {Structure of the key species in the enzymatic oxidation of methane to methanol},
+ journal = {Nature}
 }
-
+    
 @article{John:2019a,
-	Author = {{John}, C.~W. and {Proshlyakov}, D.~A.},
-	Journal = {Analytical Chemistry},
-	Title = {{FTIR Spectro-Voltammetry and Quantitative Modeling of Analytes in Kinetically Constrained Redox Mixtures}},
-	Year = 2019,
-	doi = {10.1021/acs.analchem.9b00859},
-	adsurl = {https://pubs.acs.org/doi/10.1021/acs.analchem.9b00859}
+  doi = {10.1021/acs.analchem.9b00859},
+  year = 2019,
+  publisher = {American Chemical Society ({ACS})},
+  volume = {91},
+  number = {15},
+  pages = {9563--9570},
+  author = {Christopher W. John and Denis A. Proshlyakov},
+  title = {Fourier Transform Infrared Spectrovoltammetry and Quantitative Modeling of Analytes in Kinetically Constrained Redox Mixtures},
+  journal = {Analytical Chemistry}
 }
 
 @article{John:2019b,
-	Author = {{John}, C.~W., and {Swain}, G.~M. and {Hausinger}, R.~P. and {Proshlyakov}, D.A.},
-	Journal = {Journal of Physical Chemistry B},
-	Title = {{Strongly coupled redox-linked conformational switching at the active site of the non-heme iron-dependent dioxygenase, TauD}},
-	Year = 2019,
-	doi = {10.1021/acs.jpcb.9b05866},
-	adsurl = {https://pubs.acs.org/doi/10.1021/acs.jpcb.9b05866}
[…]
