### 2019-08-27 5a3e4f4
correcting the paper
minor style changes
