### 2019-08-27 af99924
correcting the paper
minor style changes
