### 2019-08-27 687ae27
Update paper.md
Copyedits throughout the paper
